Refactor extends and stub preset utilities
