+++ packages/skeleton/scripts/compile-css-to-js.ts
@@ -4,7 +4,7 @@ import { readFileSync } from 'fs';
 import postcssImport from 'postcss-import';
 import tw from 'tailwindcss';
 import type { Config } from 'tailwindcss';
-import { corePlugin } from '../src/plugin/extends/core.js';
+import { corePlugin } from '../src/plugin/core.js';
 
 // tailwind needs to fix their exports :]
 const tailwindcss = tw as unknown as PluginCreator<string | Config | { config: string | Config }>;
