+++ packages/skeleton/src/plugin/core.ts
@@ -6,15 +6,16 @@
 import plugin from 'tailwindcss/plugin.js';
 
 // Skeleton Theme Modules
-import themeColors from './colors.js';
+import themeColors from './extends/colors.js';
 // Skeleton Design Token Modules
-import tokensBackgrounds from './tokens/backgrounds.js';
-import tokensBorders from './tokens/borders.js';
-import tokensBorderRadius from './tokens/border-radius.js';
-import tokensFills from './tokens/fills.js';
-import tokensText from './tokens/text.js';
-import tokensRings from './tokens/rings.js';
+import tokensBackgrounds from './extends/backgrounds.js';
+import tokensBorders from './extends/borders.js';
+import tokensBorderRadius from './extends/border-radius.js';
+import tokensFills from './extends/fills.js';
+import tokensText from './extends/text.js';
+import tokensRings from './extends/rings.js';
 
+// IMPORTANT: KEEP THIS FOR `/utilities/presets.ts`
 export const coreUtilities = {
 	// Implement Skeleton design token classes
 	...tokensBackgrounds(),
