+++ packages/skeleton/src/plugin/extends/backgrounds.ts
