+++ packages/skeleton/src/plugin/extends/rings.ts
