+++ packages/skeleton/src/plugin/settings.ts
