+++ packages/skeleton/src/plugin/utilities/presets.ts
@@ -1,0 +1,7 @@
+/* 
+    TODO:
+    - Translate `/components/variants.css` to Tailwind utility classes.
+    - These will be renamed from `variants` -> `presets` to avoid confusion with `@tailwind variants`
+    - Convert from CSS format to JS format
+    - Integrated into the plugin via: `core.ts` -> `coreUtilities`
+*/
