+++ packages/skeleton/tsup.config.ts
@@ -7,7 +7,7 @@ const base: Options = {
 	// treeshake: false,
 	// noExternal: [],
 	// skipNodeModulesBundle: true,
-	ignoreWatch: ['src/plugin/extends/generated']
+	ignoreWatch: ['src/plugin/generated']
 };
 
 export default defineConfig([
